+++ t3-server/src/T3/Server/MatchLogger.hs
@@ -1,0 +1,8 @@
+module T3.Server.MatchLogger
+  ( MatchLogger(..)
+  ) where
+
+import T3.Core (Result, Action, Board)
+
+class Monad m => MatchLogger m where
+  logMatch :: [Action] -> Board -> Result -> m ()
