+++ t3-server/src/T3/Server/OnTimeout.hs
@@ -1,0 +1,6 @@
+module T3.Server.OnTimeout
+  ( OnTimeout(..)
+  ) where
+
+class Monad m => OnTimeout m where
+  onTimeout :: m a -> m () -> m a
